refactor: extract click count logic
