+++ src/tab/exportProcessor/scraper/ExtractorPerplexity.js
@@ -1,7 +1,7 @@
 import {Extractor} from "./Extractor";
 import {formatLineBreaks} from "../formatter/formatText";
 import {formatLink, initTurndown, setFileHeader, turndownConverter} from "../formatter/formatMarkdown";
-import {sleep} from "../../../common/utils";
+import {sleep} from "../../../common/utils/utils";
 
 export default class ExtractorPerplexity extends Extractor {
   async extractPage(format) {
