+++ src/tab/uiEnhancer/phind/interact.js
@@ -1,5 +1,5 @@
 import {logWaitList} from "../../utils/consoleMessages";
-import {sleep} from "../../../common/utils";
+import {sleep} from "../../../common/utils/utils";
 
 export async function clickOnListElt(index, selector = '.table-responsive tr') {
   let list = document.querySelectorAll(selector);
